feat(logging): breaks up stages and glove modules * surfaces extension logs to python logger
