+++ src/crucyble/types.py
@@ -1,0 +1,4 @@
+from typing import Union
+
+def EnumUnion(enum_type):
+    return Union[int, enum_type]
